+++ src/Input/Path.hpp
@@ -27,40 +27,40 @@ class Path
 public:
     Path():_begin(0),_len(0){}
     
-    Path(unsigned int, unsigned int);
+    Path(size_t, size_t);
     
 
     // return the left bound of the interval defined by this Path (in samples)
-    unsigned int begin() const;
+    size_t begin() const;
 
     // return the right bound of the interval defined by this Path (in samples)
     // WARNING: the right bound is the last position of the interval +1
     // i.e. it does not belong to the interval
-    unsigned int end() const;
+    size_t end() const;
 
     // return the length of the interval defined by this Path (in samples)
-    unsigned int length() const;
+    size_t length() const;
 
     // the position given belongs to the interval defined by this Path
-    bool member(unsigned int);
+    bool member(size_t);
 
     // the position given belongs to the interval defined by this Path
     // and is closer to the left bound than to the right bound
-    bool aligned(unsigned int);
+    bool aligned(size_t);
     
     // sub(n, i) return a new path defined as the ith subsegment
     // of a division of this Path into n segments
     // the interval length must be divisible by n
     // UNUSED? TODO REM.
-    Path* sub(unsigned int, unsigned int);
+    Path* sub(size_t, size_t);
     
 protected:
     
     // position of left bound of interval (in samples)
-    unsigned int _begin;
+    size_t _begin;
     
     // length of interval (in samples)
-    unsigned int _len;
+    size_t _len;
     
     // _dur : inverse of duration defined by path
 };
@@ -91,18 +91,18 @@ public:
     //int date(size_t) const;
     
     // number of points of segment in the fist half of this interval
-    unsigned int l_size() const { return _seg_llen; }
+    size_t l_size() const { return _seg_llen; }
     
     // index in segment of the first element of segment inside the fist half of this interval
     // out_of_range (= size of segment) id lsize() == 0
-    unsigned int l_first() const { return _seg_lbeg; }
+    size_t l_first() const { return _seg_lbeg; }
 
     // number of points of segment in the second half of this interval
-    unsigned int r_size() const { return _seg_rlen; }
+    size_t r_size() const { return _seg_rlen; }
     
     // index in segment of the first element of segment inside the second half of this interval
     // out_of_range (= size of segment) id rsize() == 0
-    unsigned int r_first() const { return _seg_rbeg; }
+    size_t r_first() const { return _seg_rbeg; }
     
     bool habited() const { return ((_seg_llen + _seg_rlen) > 0); }
     
@@ -110,40 +110,40 @@ public:
     // division of this Path into k segments.
     // k must be > 1.
     // the interval length must be divisible by k.
-    vector<Alignment*> subs(unsigned int) const;
+    vector<Alignment*> subs(size_t) const;
     
 private:
 
     // resolution
-    unsigned int _res;  // SUPPR (deja dans segment)
+    size_t _res;  // SUPPR (deja dans segment)
     
     Segment _seg;
     
     // number of segment points in the first half of this path
-    unsigned int _seg_llen;
+    size_t _seg_llen;
 
     // index of first segment point in the first half of this path
     // or out_of_bound (segment size) is there are none
-    unsigned int _seg_lbeg;
+    size_t _seg_lbeg;
     
     // number of segment points in the second half of this path
-    unsigned int _seg_rlen;
+    size_t _seg_rlen;
 
     // index of first segment point in the second half of this path
     // or out_of_bound (segment size) is there are none
-    unsigned int _seg_rbeg;
+    size_t _seg_rbeg;
     
     
     // align(b) set the above values, starting from index b
     // and return the next index in segment to be processed
     // or the size of segment if end of segment is reached.
-    unsigned int align(unsigned int);
+    size_t align(size_t);
     
     // Alignment(s, b, l) = alignement of the interval [b..b+l] subset [0,..,r]
     // to the input segment s.
     // WARNING: must be aligned afterwards.
     // use only internaly construction of recursive paths.
-    Alignment(const Segment&, unsigned int, unsigned int);
+    Alignment(const Segment&, size_t, size_t);
     
 };
 
@@ -167,13 +167,13 @@ public:
     
     Alignment* root;
 
-    vector<AlignmentTree*> children(int);
+    vector<AlignmentTree*> children(size_t);
     
 private:
     
     // every entry in this map associate to
     // an arity n a partition a1,...,an of the root Alignment
-    map<int, vector<AlignmentTree*>> _children;
+    map<size_t, vector<AlignmentTree*>> _children;
     
 };
